tests/wpt/tests/html/canvas/offscreen/drawing-rectangles-to-the-canvas/2d.strokeRect.transform.html

<!DOCTYPE html>
<!-- DO NOT EDIT! This test has been generated by /html/canvas/tools/gentest.py. -->
<meta charset="UTF-8">
<title>OffscreenCanvas test: 2d.strokeRect.transform</title>
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/html/canvas/resources/canvas-tests.js"></script>

<h1>2d.strokeRect.transform</h1>
<p class="desc">fillRect is affected by transforms</p>


<script>
var t = async_test("fillRect is affected by transforms");
var t_pass = t.done.bind(t);
var t_fail = t.step_func(function(reason) {
    throw reason;
});
t.step(function() {

  var canvas = new OffscreenCanvas(100, 50);
  var ctx = canvas.getContext('2d');

  ctx.scale(10, 10);
  ctx.translate(0, 5);
  ctx.strokeStyle = '#0f0';
  ctx.lineWidth = 5;
  ctx.strokeRect(2.5, -2.6, 5, 0.2);
  _assertPixel(canvas, 50,25, 0,255,0,255);
  t.done();

});
</script>
