LayoutTests/imported/w3c/web-platform-tests/html/canvas/offscreen/drawing-rectangles-to-the-canvas/2d.strokeRect.transform-expected.txt

2d.strokeRect.transform

fillRect is affected by transforms


PASS fillRect is affected by transforms
